update StatusBox to MUI design
