+++ client/src/styles/__global.scss
@@ -2,6 +2,20 @@
 
 $lightblue: #a6d8d9;
 $darkblue: #1c3454;
+// src/styles/_variables.scss
+
+$background-paper: #ffffff;
+$box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+$text-primary: #000000;
+$font-weight-bold: 700;
+$background-default: #f5f5f5;
+$text-secondary: #757575;
+$font-size-medium: 16px;
+$divider: #e0e0e0;
+$border-radius: 4px;
+$primary-main: #1976d2;
+$primary-contrastText: #ffffff;
+$font-weight-medium: 500;
 
 // // global components used throughout the application
 // .container {
