+++ client/types/index.d.ts
@@ -7,3 +7,4 @@ declare module '*.jpg' {
   const value: string;
   export default value;
 }
+
